requirements.txt
alabaster==0.7.12
annotated-types==0.7.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
black==25.1.0
blinker==1.4
breezy==3.2.1
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.0
cftime==1.5.2
chardet==4.0.0
chex==0.1.89
click==8.0.3
colorama==0.4.4
colorcet==2.0.2
commonmark==0.9.1
configmypy==0.2.0
configobj==5.0.6
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
dnspython==2.1.0
docker-pycreds==0.4.0
docutils==0.17.1
dulwich==0.20.31
empy==3.3.4
etils==1.12.2
exceptiongroup==1.2.2
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
fenics==2019.1.0
filelock==3.13.1
flatbuffers==25.2.10
flax==0.3.3
fonttools==4.56.0
fs==2.4.12
fsspec==2024.6.1
future==0.18.2
gast==0.6.0
genpy==0.6.16
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gpg==1.16.0
grpcio==1.71.0
gyp==0.1
h5py==3.13.0
h5py.-debian-h5py-serial==3.6.0
html5lib==1.1
httplib2==0.20.2
humanize==4.12.1
idna==3.3
imageio==2.4.1
imagesize==1.3.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
iniconfig==1.1.1
jax==0.2.20
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.1.75
jeepney==0.7.1
Jinja2==3.0.3
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==3.9.0
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
matplotlib==3.10.1
meshio==5.3.0
message-filters==1.15.14
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
msgpack==1.1.0
mshr==2019.2.0.dev0
mypy-extensions==1.0.0
namex==0.0.8
nanobind==2.5.0
nest-asyncio==1.6.0
netCDF4==1.5.8
netifaces==0.11.0
networkx==3.3
neuraloperator==0.3.0
notify2==0.3
numba==0.55.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
packaging==24.2
param===None
paramiko==2.9.3
pathspec==0.12.1
patiencediff==0.2.1
petsc4py==3.15.1
pexpect==4.8.0
pillow==11.1.0
pkgconfig==1.5.5
platformdirs==4.3.7
pluggy==0.13.0
ply==3.11
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pusimp==0.1.0
py==1.10.0
pybind11==2.9.1
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pyct==0.4.7a3
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
pytest-mock==3.14.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
python-gnupg==0.4.8
python-qt-binding==0.4.4
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
rich==11.2.0
roman==3.3
rosbag==1.15.14
rosclean==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslz4==1.15.14
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.3.0
rospy==1.15.14
rosservice==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit_build_core==0.11.0
scipy==1.12.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sensor-msgs==1.13.1
sentry-sdk==2.24.0
setproctitle==1.3.5
simplejson==3.20.1
six==1.17.0
slepc4py==3.15.1
smmap==5.0.2
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
ssh-import-id==5.11
sympy==1.13.1
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorly==0.9.0
tensorly-torch==0.5.0
tensorstore==0.1.72
termcolor==2.5.0
tf==1.13.2
tf2-py==0.7.5
tf2-ros==0.7.5
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.6.0+cu126
torch-scatter==2.1.2
torch_harmonics==0.7.4
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
treescope==0.1.9
triton==3.2.0
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
UNKNOWN==0.0.0
urllib3==2.3.0
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wandb==0.19.8
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.13.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
zipp==1.0.0
